refactor(application): move input handling to InputView class
- Extract user input logic from application layer into dedicated InputView
